refactor: config constructor separated
