+++ src/services/config/mod.rs
@@ -1,1 +1,2 @@
 pub mod config_actor;
+mod constructor;
